+++ src/components/StoreComponent.ts
@@ -1,16 +1,18 @@
 /* eslint-disable no-restricted-globals */
-import {routerMiddleware, connectRouter, RouterState} from 'connected-react-router';
+import {createRouterMiddleware, createRouterReducer, ReduxRouterState} from '@lagunovsky/redux-react-router';
 import {
     History,
     createBrowserHistory,
     createMemoryHistory,
     createHashHistory,
-    LocationState,
 } from 'history';
 import _get from 'lodash-es/get';
 import _isPlainObject from 'lodash-es/isPlainObject';
 import _merge from 'lodash-es/merge';
-import {createStore, applyMiddleware, compose, Store, Dispatch, Unsubscribe, Reducer, AnyAction} from 'redux';
+import {
+    createStore, applyMiddleware, compose, Store, Dispatch, Unsubscribe, Reducer, UnknownAction, MiddlewareAPI,
+    StoreEnhancer,
+} from 'redux';
 
 import {IComponents} from '../providers/ComponentsProvider';
 
@@ -27,17 +29,17 @@ interface IStoreComponentConfig {
 }
 
 type AsyncReducersMap = {
-    router?: Reducer<RouterState<LocationState>, AnyAction>,
-    [key: string]: Reducer<any, AnyAction> | undefined,
+    router?: Reducer<ReduxRouterState, UnknownAction>,
+    [key: string]: Reducer<any, UnknownAction> | undefined,
 };
 
 type RootState<R extends AsyncReducersMap> = {
-    [K in keyof R]: R[K] extends Reducer<infer S, AnyAction> ? S : never;
+    [K in keyof R]: R[K] extends Reducer<infer S, UnknownAction> ? S : never;
 };
 
 type CreateRootReducer = <R extends AsyncReducersMap>(
     reducers: R,
-) => Reducer<RootState<R>, AnyAction>;
+) => Reducer<RootState<R>, UnknownAction>;
 
 export interface IStoreComponent {
     /**
@@ -109,7 +111,7 @@ export default class StoreComponent implements IStoreComponent {
 
     reducers: CreateRootReducer;
 
-    _routerReducer: Reducer<RouterState<LocationState>>;
+    _routerReducer: Reducer<ReduxRouterState>;
 
     history: History | null;
 
@@ -165,12 +167,7 @@ export default class StoreComponent implements IStoreComponent {
                 ..._get(initialState, 'config.store.history', {}),
                 ...config.history,
             });
-
-            // Add '?' for fix connected-react-router
-            if (process.env.IS_SSR && !this.history.location.search) {
-                this.history.location.search = '?';
-            }
-            this._routerReducer = connectRouter(this.history);
+            this._routerReducer = createRouterReducer(this.history);
         }
 
         if (!this.store) {
@@ -182,12 +179,18 @@ export default class StoreComponent implements IStoreComponent {
                 ),
                 initialState,
                 compose(
-                    applyMiddleware(({getState}) => next => action => this._prepare(action, next, getState)),
-                    applyMiddleware(routerMiddleware(this.history)),
+                    applyMiddleware(
+                        ({getState}: MiddlewareAPI) => (next: Dispatch) => (action: unknown) => this._prepare(
+                            action,
+                            next,
+                            getState,
+                        ),
+                    ),
+                    applyMiddleware(createRouterMiddleware(this.history)),
                     !process.env.IS_SSR && window.__REDUX_DEVTOOLS_EXTENSION__ && process.env.PLATFORM !== 'mobile'
                         ? window.__REDUX_DEVTOOLS_EXTENSION__()
                         : f => f,
-                ),
+                ) as StoreEnhancer,
             );
         }
     }